Allow bin_path to be specified via env variable
This solves a problem where the default binary found by `which` is the Ruby gem
version of `terminal-notifier`. This version is slow, especially in a Bundler
environment.

In this scenario we would much rather use the Homebrew-installed version at
`/usr/local/bin/terminal-notifier` which is noticeable faster.

This commit allows us to set an environment variable to explicitly specify the
binary to use, like this:

export TERMINAL_NOTIFIER_BIN=/usr/local/bin/terminal-notifier